warnings fixed

+++ libraries/Elegoo_TFTLCD/pin_magic.h
@@ -163,13 +163,13 @@
     DDRH &= ~B01111000; DDRB &= ~B10110000; DDRG &= ~B00100000; }
 
  #else // Mega w/Breakout board
-		#define write8inline(d) {\                        
+		#define write8inline(d) {\
 		PORTH &= ~(0x78);\
   	PORTH |= ((d&0xC0) >> 3) | ((d&0x3) << 5);\
   	PORTE &= ~(0x38);\
   	PORTE |= ((d & 0xC) << 2) | ((d & 0x20) >> 2);\
   	PORTG &= ~(0x20);\
-  	PORTG |= (d & 0x10) << 1; \ 
+  	PORTG |= (d & 0x10) << 1; \
     WR_STROBE; }
   #define read8inline(result) { RD_ACTIVE; DELAY7; result = (PINH & 0x60) >> 5;result |= (PINH & 0x18) << 3;result |= (PINE & 0x8) << 2;result |= (PINE & 0x30) >> 2;result |= (PING & 0x20) >> 1;RD_IDLE;}
   #define setWriteDirInline() { DDRH |= 0x78;DDRE |= 0x38;DDRG |= 0x20; }
@@ -284,7 +284,7 @@
    PIO_Clear(PIOB, (((~d) & 0x20)<<(27-5))); \
    WR_STROBE; }
 
-  #define read8inline(result) { \    
+  #define read8inline(result) { \
    RD_ACTIVE;   \
    delayMicroseconds(1);      \
    result = (((PIOC->PIO_PDSR & (1<<23)) >> (23-7)) | ((PIOC->PIO_PDSR & (1<<24)) >> (24-6)) | \
@@ -403,7 +403,7 @@
 #define writeRegister16inline(a, d) { \
   uint8_t hi, lo; \
   hi = (a) >> 8; lo = (a); CD_COMMAND; write8(hi); write8(lo); \
-  hi = (d) >> 8; lo = (d); CD_DATA   ; write8(hi); write8(lo); }
+  hi = (d) >> 8; lo = (uint8_t)(d); CD_DATA   ; write8(hi); write8(lo); }
 
 // Set value of 2 TFT registers: Two 8-bit addresses (hi & lo), 16-bit value
 #define writeRegisterPairInline(aH, aL, d) { \
